+++ Rh-Backend/Services/FeriasService.cs
@@ -1,5 +1,6 @@
 
 using AutoMapper;
+using Microsoft.AspNetCore.Http.HttpResults;
 using Rh_Backend.DTO;
 using Rh_Backend.Exceptions;
 using Rh_Backend.Models;
@@ -55,6 +56,7 @@ namespace Rh_Backend.Services
             try
             {
                 if (idFuncionario <= 0) throw new BadRequestException("Id deve ser maior que 0");
+                if (!await _funcionarioRepository.ExistsAsync(idFuncionario)) throw new NotFoundException("Funcionário não existe");
                 var ferias = await _feriasRepository.GetByIdFuncionarioAsync(idFuncionario) ?? throw new NotFoundException("Nenhum registro de férias encontrado para o funcionário.");
                 return _mapper.Map<IEnumerable<FeriasReadDTO>>(ferias);
             }
@@ -84,7 +86,9 @@ namespace Rh_Backend.Services
         {
             try
             {
+                Console.WriteLine($"Status recebido: '{status}'");
                 if (status.Any(char.IsDigit)) throw new BadRequestException("Apenas letras são permitidas");
+                if (!(status.Equals("Em andamento") || status.Equals("Pendente") || status.Equals("Concluídas"))) throw new BadRequestException("Status inválido");
                 var ferias = await _feriasRepository.GetByStatusAsync(status) ?? throw new Exception("Nenhum registro de férias encontrado.");
                 return _mapper.Map<IEnumerable<FeriasReadDTO>>(ferias);
             }
@@ -100,6 +104,8 @@ namespace Rh_Backend.Services
             try
             {
                 if (status.Any(char.IsDigit)) throw new BadRequestException("Apenas letras são permitidas");
+                if (!(status.Equals("Em andamento") || status.Equals("Pendente") || status.Equals("Concluídas"))) throw new BadRequestException("Status inválido");
+                if (!await _funcionarioRepository.ExistsAsync(idFuncionario)) throw new BadRequestException("Funcionario não existe");
                 if (idFuncionario <= 0) throw new BadRequestException("Id não pode ser menor que 0");
                 var ferias = await _feriasRepository.GetByFuncionarioAndStatusAsync(idFuncionario, status) ?? throw new Exception("Nenhum registro de férias encontrado.");
                 return _mapper.Map<IEnumerable<FeriasReadDTO>>(ferias);
@@ -122,7 +128,7 @@ namespace Rh_Backend.Services
 
                 var feriasModel = _mapper.Map<FeriasModel>(ferias);
                 if (ferias.DataInicio > DateTime.Now) feriasModel.Status = "Pendente";
-                if (DateTime.Now >= ferias.DataInicio && DateTime.Now <= ferias.DataFim) feriasModel.Status = "Em Andamento";
+                if (DateTime.Now >= ferias.DataInicio && DateTime.Now <= ferias.DataFim) feriasModel.Status = "Em andamento";
                 if (ferias.DataFim < DateTime.Now) feriasModel.Status = "Concluídas";
 
 
@@ -161,7 +167,7 @@ namespace Rh_Backend.Services
             try
             {
                 if (id <= 0) throw new BadRequestException("ID de férias inválido.");
-                var ferias = await _feriasRepository.GetByIdAsync(id) ?? throw new NotFoundException("Registro de férias não encontrado.");
+                if (!await _feriasRepository.ExistsAsync(id)) throw new NotFoundException("Registro de férias não encontrado.");
                 return await _feriasRepository.DeleteAsync(id);
             }
             catch (Exception ex)
@@ -176,7 +182,9 @@ namespace Rh_Backend.Services
             try
             {
                 if (id <= 0) throw new BadRequestException("ID de férias inválido.");
-                return await _feriasRepository.ExistsAsync(id);
+                var ferias = await _feriasRepository.ExistsAsync(id);
+                if (!ferias) throw new NotFoundException("Ferias não encontrada");
+                return ferias;
             }
             catch (Exception ex)
             {
